feat: add game engine
