+++ src/index.ts
@@ -5,6 +5,7 @@ import { WebSocketServer, createWebSocketStream, WebSocket } from 'ws';
 import { AddShipsRequest, CreateGameResponse, InputMessage, RegisterResponse, StartGameResponse, TurnResponse, UpdateRoomEvent } from './messages';
 import { StateManager } from './state-manager';
 import { AppState, GameId } from './app.state';
+import { GameEngine } from './game-engine';
 
 
 export const httpServer = http.createServer(function (req, res) {
@@ -26,92 +27,25 @@ let userCount = 0;
 try {
   const WEBSOCKET_PORT = Number(process.env.WEBSOCKET_PORT) || 3000;
   const wsServer = new WebSocketServer({ port: WEBSOCKET_PORT });
-  const stateManager = new StateManager();
+  const gameEngine = new GameEngine();
   wsServer.on('connection', (ws) => {
-    const userId = userCount++;
-    let userName: string | undefined;
-    let gameId: GameId | undefined;
+    console.log('Connection received ')
 
-
-    console.log('Connection received ', userId)
-
-    stateManager.subscribe((event, state) => {
-      console.log('App state updated', event, state)
-      if (event.type === 'room_created' || event.type === 'user_registered') {
-        const updatedRoomList = listRooms(state, userId);
-        sendMessage(updatedRoomList, ws, userId);
-      }
-      if (event.type === 'user_added_to_room') {
-        const updatedRoomList = listRooms(state, userId);
-        const game = Object.values(state.games)[0];
-        const players = Object.keys(game.players).map(Number)
-        if (players.includes(userId)) {
-          const createGame: CreateGameResponse = {
-            type: 'create_game',
-            data: {
-              idGame: game.id,
-              idPlayer: players.filter((player) => player !== userId)[0]
-            },
-            id: 0
-          }
-          ws.send(serializeMessage((createGame)))
-          gameId = game.id;
-        }
-        sendMessage(updatedRoomList, ws);
-      }
-    })
 
     const wsStream = createWebSocketStream(ws, {
       encoding: 'utf8',
       decodeStrings: false,
     });
+    let callback: (data: any) => void | undefined;
     wsStream.on('data', (data: string) => {
       const dataObj: InputMessage = deserializeMessage(data);
       console.log(`Data received: ${JSON.stringify(dataObj)}`)
       if (dataObj.type === 'reg') {
-        stateManager.publishEvent({
-          type: 'user_registered',
-          id: userId,
-          name: dataObj.data.name,
-        })
-        const registerResponse: RegisterResponse = {
-          type: 'reg',
-          data: {
-            name: dataObj.data.name,
-            index: userCount,
-            error: false,
-            errorText: ''
-          },
-          id: 0
-        }
-        userName = dataObj.data.name;
-        ws.send(serializeMessage(registerResponse));
+        callback = gameEngine.regUser(dataObj, (data) => sendMessage(data, ws))
+      } else {
+        callback?.(dataObj)
       }
-      if (dataObj.type === 'create_room') {
-        stateManager.publishEvent({
-          type: 'room_created',
-          ownerId: userId
-        })
-      }
-      if (dataObj.type === 'add_user_to_room') {
-        stateManager.publishEvent({
-          type: 'user_added_to_room',
-          roomId: dataObj.data.indexRoom,
-          userId: userId
-        })
-      }
-      if (dataObj.type === 'add_ships') {
-        if (gameId == null) {
-          console.error(`Game with ${userId} does not exist`);
-          return
-        }
-        stateManager.publishEvent({
-          type: 'ships_added',
-          gameId: gameId,
-          userId: userId,
-          ships: dataObj.data.ships,
-        });
-      }
+
     });
     wsStream.on('error', (err) => {
       console.error(err)
@@ -147,23 +81,5 @@ function deserializeMessage<T>(input: string): T {
   if (obj.data)
     obj.data = JSON.parse(obj.data);
   return obj;
-}
-
-function listRooms(state: AppState, userId: number): UpdateRoomEvent {
-  return {
-    type: 'update_room',
-    id: 0,
-    data: state.rooms.map((val, i) => {
-      const player1 = state.users.find(x => x.id === val.player1)!;
-      return {
-        roomId: val.id,
-        roomUsers: [player1]
-          .map(p => ({
-            index: p!.id,
-            name: p!.name
-          }))
-      }
-    })
-  }
 }
